requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
couchbase==4.2.1
dataclasses-json==0.6.6
distro==1.9.0
Flask==3.0.3
Flask-SocketIO==5.3.6
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.5
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langchain-tools==0.1.34
langsmith==0.1.75
MarkupSafe==2.1.5
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.32.0
orjson==3.10.3
packaging==23.2
phonenumbers==8.13.38
pydantic==1.10.8
pydantic_core==2.18.4
PyJWT==2.8.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.2
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
simple-websocket==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.30
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.4
twilio==9.1.1
typing-inspect==0.9.0
typing_extensions==4.12.1
urllib3==2.2.1
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
